Backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.6.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
frozenlist==1.4.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langsmith==0.0.41
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.1
pydantic==2.4.2
pydantic_core==2.10.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
yarl==1.9.2
